\name{ordiBreadth-package}
\alias{ordiBreadth-package}
\alias{ordiBreadth}
\docType{package}
\title{
Calculates ordinated diet breadth.
}
\description{
Calculates ordinated diet breadth and provides plotting functions.
}
\details{
\tabular{ll}{
Package: \tab ordiBreadth\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-08-07\cr
License: \tab GPL 3.0\cr
}
Users provide a matrix of host associations, where rows are consumers (e.g., heribivores) and columns are resources (e.g., plants). The function ordi.breadth calculates the ordinated breadth for all consumers in the matrix and summary.hbreadth summarizes the output and dist.group.plot can be used for visualization.  
}
\author{
James Fordyce

Maintainer: James Fordyce <jfordyce@utk.edu>
}
\references{
Fordyce, J.A., C.C. Nice, C.A. Hamm, & M.L. Forister. Quantifying diet breadth through ordination of host association. Ecology
}
\keyword{ package }
